+++ lib/oneview-sdk/resource/api200/volume_attachment.rb
@@ -47,9 +47,11 @@ module OneviewSDK
 
       # Gets the list of extra unmanaged storage volumes
       # @param [OneviewSDK::Client] client The client object for the OneView appliance
+      # @return [Array] Array of extra unmanaged storage volumes
       def self.get_extra_unmanaged_volumes(client)
         response = client.rest_get(BASE_URI + '/repair?alertFixType=ExtraUnmanagedStorageVolumes')
-        client.response_handler(response)
+        body = client.response_handler(response)
+        body['members'].map { |data| new(client, data) }
       end
 
       # Removes extra presentations from a specific server profile