+++ spec/integration/resource/api500/synergy/id_pool/update_spec.rb
@@ -1,0 +1,6 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API500::Synergy::IDPool, integration: true, type: UPDATE do
+  let(:current_client) { $client_500_synergy }
+  include_examples 'IDPoolUpdateExample', 'integration api500 context'
+end