+++ spec/unit/resource/api200/server_profile_spec.rb
@@ -90,12 +90,6 @@ RSpec.describe OneviewSDK::ServerProfile do
       @server_hardware_uri = '/rest/fake/server-hardwares/test'
     end
 
-    it 'will set the serverHardwareUri correctly' do
-      @server_hardware['uri'] = @server_hardware_uri
-      @item.set_server_hardware(@server_hardware)
-      expect(@item['serverHardwareUri']).to eq(@server_hardware_uri)
-    end
-
     it 'will retrieve and set the serverHardwareUri correctly' do
       expect(@client_200).to receive(:rest_get).with('/rest/server-hardware', {})
         .and_return(FakeResponse.new(members: [
@@ -119,12 +113,6 @@ RSpec.describe OneviewSDK::ServerProfile do
     before :each do
       @server_hardware_type = OneviewSDK::ServerHardwareType.new(@client_200, name: 'server_hardware_type')
       @server_hardware_type_uri = '/rest/fake/server-hardware-types/test'
-    end
-
-    it 'will set the serverHardwareTypeUri correctly' do
-      @server_hardware_type['uri'] = @server_hardware_type_uri
-      @item.set_server_hardware_type(@server_hardware_type)
-      expect(@item['serverHardwareTypeUri']).to eq(@server_hardware_type_uri)
     end
 
     it 'will retrieve and set the serverHardwareTypeUri correctly' do
@@ -152,12 +140,6 @@ RSpec.describe OneviewSDK::ServerProfile do
       @enclosure_group_uri = '/rest/fake/enclosure-groups/test'
     end
 
-    it 'will set the enclosureGroupUri correctly' do
-      @enclosure_group['uri'] = @enclosure_group_uri
-      @item.set_enclosure_group(@enclosure_group)
-      expect(@item['enclosureGroupUri']).to eq(@enclosure_group_uri)
-    end
-
     it 'will retrieve and set the enclosureGroupUri correctly' do
       expect(@client_200).to receive(:rest_get).with('/rest/enclosure-groups', {})
         .and_return(FakeResponse.new(members: [
@@ -181,12 +163,6 @@ RSpec.describe OneviewSDK::ServerProfile do
     before :each do
       @enclosure = OneviewSDK::Enclosure.new(@client_200, name: 'enclosure')
       @enclosure_uri = '/rest/fake/enclosures/test'
-    end
-
-    it 'will set the enclosureUri correctly' do
-      @enclosure['uri'] = @enclosure_uri
-      @item.set_enclosure(@enclosure)
-      expect(@item['enclosureUri']).to eq(@enclosure_uri)
     end
 
     it 'will retrieve and set the enclosureUri correctly' do
@@ -276,10 +252,9 @@ RSpec.describe OneviewSDK::ServerProfile do
 
   describe '#self.get_available_storage_systems' do
     it 'retrieves available storage system based on a query' do
-      expect(@client_200).to receive(:rest_get)
-        .with("#{OneviewSDK::ServerProfile::BASE_URI}/available-storage-systems")
-        .and_return(FakeResponse.new('it' => 'works'))
-      expect(OneviewSDK::ServerProfile.get_available_storage_systems(@client_200)['it']).to eq('works')
+      fake_response = FakeResponse.new(members: [{ it: 'works' }])
+      allow(@client_200).to receive(:rest_get).and_return(fake_response)
+      expect(OneviewSDK::ServerProfile.get_available_storage_systems(@client_200).first['it']).to eq('works')
     end
   end
 
@@ -327,13 +302,13 @@ RSpec.describe OneviewSDK::ServerProfile do
 
   describe '#update_from_template' do
     it 'transforms an existing profile' do
-      patch_ops = [{ 'op' => 'replace', 'path' => '/templateCompliance', 'value' => 'Compliant' }]
+      patch_ops = [{ op: 'replace', path: '/templateCompliance', value: 'Compliant' }]
       request = {
         'If-Match' => @item['eTag'],
         'body' => patch_ops
       }
       expect(@client_200).to receive(:rest_patch)
-        .with(@item['uri'], request)
+        .with(@item['uri'], request, 200)
         .and_return(FakeResponse.new)
       expect { @item.update_from_template }.to_not raise_error
     end
@@ -453,6 +428,8 @@ RSpec.describe OneviewSDK::ServerProfile do
   describe 'Volume attachment operations' do
     it 'can call the #add_volume_attachment using a specific already created Volume' do
       options = { uri: '/fake/volume', storagePoolUri: '/fake/storage-pool', storageSystemUri: '/fake/storage-system' }
+      fake_response = FakeResponse.new(members: [options])
+      expect(@client_200).to receive(:rest_get).with('/rest/storage-volumes').and_return(fake_response)
       volume = OneviewSDK::Volume.new(@client_200, options)
       @item.add_volume_attachment(volume)
 
@@ -466,6 +443,8 @@ RSpec.describe OneviewSDK::ServerProfile do
     describe 'can call #remove_volume_attachment' do
       it 'and remove attachment with id 0' do
         options = { uri: '/fake/volume', storagePoolUri: '/fake/storage-pool', storageSystemUri: '/fake/storage-system' }
+        fake_response = FakeResponse.new(members: [options])
+        expect(@client_200).to receive(:rest_get).with('/rest/storage-volumes').and_return(fake_response)
         volume = OneviewSDK::Volume.new(@client_200, options)
         @item.add_volume_attachment(volume, 'id' => 7)
         expect(@item['sanStorage']['volumeAttachments'].size).to eq(1)
@@ -478,6 +457,8 @@ RSpec.describe OneviewSDK::ServerProfile do
 
       it 'and return nil if no attachment found' do
         options = { uri: '/fake/volume', storagePoolUri: '/fake/storage-pool', storageSystemUri: '/fake/storage-system' }
+        fake_response = FakeResponse.new(members: [options])
+        expect(@client_200).to receive(:rest_get).with('/rest/storage-volumes').and_return(fake_response)
         volume = OneviewSDK::Volume.new(@client_200, options)
         @item.add_volume_attachment(volume, 'id' => 7)
         expect(@item['sanStorage']['volumeAttachments'].size).to eq(1)
@@ -495,13 +476,16 @@ RSpec.describe OneviewSDK::ServerProfile do
     end
 
     it 'can call #create_volume_with_attachment and generate the data required for a new Volume with attachment' do
-      storage_pool = OneviewSDK::StoragePool.new(@client_200, uri: 'fake/storage-pool')
+      options = { uri: 'fake/storage-pool', storageSystemUri: 'fake/storage-system' }
+      fake_response = FakeResponse.new(members: [options])
+      expect(@client_200).to receive(:rest_get).with('/rest/storage-pools').and_return(fake_response)
+
+      storage_pool = OneviewSDK::StoragePool.new(@client_200, options)
       volume_options = {
         name: 'TestVolume',
         description: 'Test Volume for Server Profile Volume Attachment',
         provisioningParameters: {
           provisionType: 'Full',
-          shareable: true,
           requestedCapacity: 1024 * 1024 * 1024
         }
       }
@@ -511,7 +495,7 @@ RSpec.describe OneviewSDK::ServerProfile do
       expect(va['volumeUri']).not_to be
       expect(va['volumeStorageSystemUri']).not_to be
       expect(va['volumeStoragePoolUri']).to eq('fake/storage-pool')
-      expect(va['volumeShareable']).to eq(true)
+      expect(va['volumeShareable']).to eq(false)
       expect(va['volumeProvisionedCapacityBytes']).to eq(1024 * 1024 * 1024)
       expect(va['volumeProvisionType']).to eq('Full')
     end